+++ src/models.ts
@@ -24,6 +24,7 @@ export interface Settings {
   topperOverflow: boolean;
   topperOverflowOffset: number;
   lowerThanBottom: boolean;
+  upperThanTop: boolean;
   showDraggable: boolean;
   draggableOver: boolean;
   clickBottomOpen: boolean;